transformers/encoder_only_transformers/version_1/requirements.txt
tiktoken
numpy
langchain_community
tiktoken
beautifulsoup4
pypdf
lxml
torch
torchdata
transformers>=4.34
accelerate
evaluate
rouge_score
loralib
peft
datasets
bitsandbytes
torchmetrics
hydra-core
huggingface-hub
wandb
seaborn
sklearn